# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DataMart GH reseller API proxy
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /order/{reference}:
    get:
      operationId: getOrderStatus
      tags: [orders]
      summary: Get order status
      description: Get the status of a specific order by reference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStatusResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /withdrawals:
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create a withdrawal
      description: Withdraw funds from the wallet to a mobile money number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "200":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{reference}:
    get:
      operationId: getWithdrawalStatus
      tags: [withdrawals]
      summary: Get withdrawal status
      description: Check the status of a withdrawal by reference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Withdrawal status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalStatusResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OrderData:
      type: object
      properties:
        orderId:
          type: string
        reference:
          type: string
        phoneNumber:
          type: string
        network:
          type: string
        capacity:
          type: number
        price:
          type: number
        orderStatus:
          type: string
        processingMethod:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - orderId
        - reference
        - phoneNumber
        - network
        - capacity
        - price
        - orderStatus
    OrderStatusResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: "#/components/schemas/OrderData"
      required:
        - status
...
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
        phoneNumber:
          type: string
        network:
          type: string
          enum: [MTN, TELECEL, AIRTELTIGO]
        clientRef:
          type: ["string", "null"]
      required:
        - amount
        - phoneNumber
        - network
    WithdrawalData:
      type: object
      properties:
        reference:
          type: string
        amount:
          type: number
        fee:
          type: number
        phoneNumber:
          type: string
        network:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - reference
        - amount
        - phoneNumber
        - network
        - status
    WithdrawalResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/WithdrawalData"
      required:
        - status
        - message
    WithdrawalErrorResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        code:
          type: ["string", "null"]
      required:
        - status
        - message
    WithdrawalStatusData:
      type: object
      properties:
        reference:
          type: string
        amount:
          type: number
        fee:
          type: number
        phoneNumber:
          type: string
        network:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - reference
        - amount
        - phoneNumber
        - network
        - status
    WithdrawalStatusResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: "#/components/schemas/WithdrawalStatusData"
      required:
        - status